--- buildSrc/build.gradle.kts
@@ -1,7 +1,0 @@
-plugins {
-    `kotlin-dsl`
-}
-
-repositories {
-    mavenCentral()
-}
